Security/Upload Draft to Firebase
Added in user and admin security checks, aswell as limiting calls to api's that are adjustable in firebase. Likely would need to be on back end to be fully effective however it works well enough for now as a feature. It would have had to be implemented early on for ease